Splitting IRPCIterator.h into two files to avoid the use of @stubgen:skip(because the stubgenerator can not parse the content of IteratorType.h)
