+++ Source/com/IIteratorType.h
@@ -2,7 +2,7 @@
  * If not stated otherwise in this file or this component's LICENSE file the
  * following copyright and licenses apply:
  *
- * Copyright 2020 Metrological
+ * Copyright 2021 Metrological
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,22 +18,28 @@
  */
 
 #pragma once
+#include <core/core.h>
 
-#ifndef MODULE_NAME
-#error "Please define a MODULE_NAME that describes the binary/library you are building."
-#endif
+namespace WPEFramework {
 
-#include "Administrator.h"
-#include "Communicator.h"
-#include "ConnectorType.h"
-#include "IteratorType.h"
-#include "IIteratorType.h"
-#include "IStringIterator.h"
-#include "IUnknown.h"
-#include "IValueIterator.h"
-#include "ICOM.h"
-#include "Messages.h"
+namespace RPC {
 
-#if defined(__WINDOWS__) && !defined(COM_EXPORTS)
-#pragma comment(lib, "com.lib")
-#endif
+    // @iterator
+    template<typename ELEMENT, const uint32_t INTERFACE_ID>
+    struct IIteratorType : virtual public Core::IUnknown {
+
+        typedef ELEMENT Element;
+
+        enum { ID = INTERFACE_ID };
+
+        virtual ~IIteratorType(){};
+
+        virtual bool Next(ELEMENT& info /* @out */) = 0;
+        virtual bool Previous(ELEMENT& info /* @out */) = 0;
+        virtual void Reset(const uint32_t position) = 0;
+        virtual bool IsValid() const = 0;
+        virtual uint32_t Count() const = 0;
+        virtual ELEMENT Current() const = 0;
+    };
+}
+}
